# Speed up py-dcqc test suite