--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -65,25 +65,18 @@ jobs:
         - macos-latest
         # TODO: Debug the Windows issues
         # - windows-latest
+    env:
+      OS: ${{ matrix.platform }}
+      PYTHON: ${{ matrix.python }}
     runs-on: ${{ matrix.platform }}
     steps:
       - uses: actions/checkout@v3
       - uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python }}
-      - name: Retrieve pre-built distribution files
-        uses: actions/download-artifact@v3
+      - uses: actions/download-artifact@v3
         with: {name: python-distribution-files, path: dist/}
-      - name: Run tests (without integration tests)
-        if: matrix.platform != 'ubuntu-latest' || matrix.python != '3.11'
-        env:
-          SYNAPSE_AUTH_TOKEN: ${{ secrets.SYNAPSE_AUTH_TOKEN }}
-        run: >-
-          pipx run --spec 'tox~=3.0' tox
-          --installpkg '${{ needs.prepare.outputs.wheel-path }}'
-          -- -rFEx --durations 10 --color yes -m "not integration"
-      - name: Run tests (with integration tests)
-        if: matrix.platform == 'ubuntu-latest' && matrix.python == '3.11'
+      - name: Run tests
         env:
           SYNAPSE_AUTH_TOKEN: ${{ secrets.SYNAPSE_AUTH_TOKEN }}
         run: >-
@@ -92,13 +85,14 @@ jobs:
           -- -rFEx --durations 10 --color yes
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v3
-        if: matrix.platform == 'ubuntu-latest' && matrix.python == '3.11'
         with:
-          files: coverage.xml
-          name: ${{ matrix.platform }} - py${{ matrix.python }}
-          verbose: true
           # CodeCov can be flaky, so this step is not required for success
           fail_ci_if_error: false
+          files: coverage.xml
+          # Using matrix pattern from `codecov/codecov-action` README:
+          # https://github.com/codecov/codecov-action#example-workflowyml-with-codecov-action
+          env_vars: OS,PYTHON
+          verbose: true
 
   pypi-publish:
     needs: [prepare, test]