--- a/setup.cfg
+++ b/setup.cfg
@@ -109,6 +109,7 @@ console_scripts =
 #          Comment those flags to avoid this pytest issue.
 addopts =
     --cov dcqc --cov-report term-missing --cov-report xml
+    -m "not slow"
     --verbose
 norecursedirs =
     dist
@@ -121,9 +122,9 @@ testpaths =
     tests
     demos
 # Use pytest markers to select/deselect specific tests
-# markers =
-#     slow: mark tests as slow (deselect with '-m "not slow"')
-#     system: mark end-to-end system tests
+markers =
+    slow: mark tests as slow (deselect with '-m "not slow"')
+    acceptance: mark end-to-end acceptance tests
 
 [devpi:upload]
 # Options for the devpi: PyPI server and packaging tool