--- a/setup.cfg
+++ b/setup.cfg
@@ -124,7 +124,7 @@ testpaths =
 # Use pytest markers to select/deselect specific tests
 markers =
     slow: mark tests as slow (deselect with '-m "not slow"')
-    acceptance: mark end-to-end acceptance tests
+    # acceptance: mark end-to-end acceptance tests
 
 [devpi:upload]
 # Options for the devpi: PyPI server and packaging tool