--- a/tests/test_acceptance.py
+++ b/tests/test_acceptance.py
@@ -9,7 +9,6 @@ from dcqc.utils import open_parent_fs
 
 
 @pytest.mark.slow
-@pytest.mark.acceptance
 def test_json_report_generation(get_data):
     # GIVEN a list of external tests to skip (to remain self-contained)
     all_tests = TestABC.list_subclasses()