--- a/tox.ini
+++ b/tox.ini
@@ -20,7 +20,8 @@ extras =
     testing
     all
 commands =
-    pytest {posargs}
+    # The `-m ""` overrides the `-m "not slow"` in setup.cfg
+    pytest {posargs} -m ""
 
 
 [testenv:lint]